added zoneminder adapter, issues, translogs, changes all-around

+++ server/arlo_st/library.py
@@ -1,11 +1,9 @@
 import threading
 import uuid
-from queue import simplequeue
 import requests
-from peewee import doesnotexist
-from arlo_st.db import library
+from arlo_st.db import Library
 from arlo_st import libraryFs
-
+from peewee import DoesNotExist
 
 class RecordingNotFound(Exception):
     def __init__(self, msg):
@@ -13,8 +11,8 @@ class RecordingNotFound(Exception):
     def __str__(self):
         return f'Recording not found: {self.message}'
 
-
-#simply swaps the DoesNotExist exception with RecordingNotFound (for flask handling)
+# simply swaps the DoesNotExist exception with RecordingNotFound (to
+# be handled by flask)
 def get_library_model(video_id):
     try:
         return Library.get_by_id(video_id)
@@ -22,44 +20,45 @@ def get_library_model(video_id):
         raise RecordingNotFound("recording does not exist")
 
 def do_sync_recording(library_model, file_stream):
-    local_id = str(uuid.uuidv4())
-
-    #write new file
-    libraryFs.write_file(local_id, file_stream)
-
-    #remove old file
-    libraryFs.remove_file(library_model.location_local)
-
-    #update db
-    library_model.location_local = local_id
-    library_model.save()
-        
-def snyc_recording(video_id):
-    library_model = get_library_model(video_id)
-    
+    local_id = str(uuid.uuid4())
     try:
-        remote_url = library_model.location_remote
-        r = requests.get(remote_url, stream=True, timeout=3)
-        r.raise_for_status()
-        
-        q = SimpleQueue()
-        
-        threading.Thread(
-            target=do_sync_recording,
-            args=(library_model, r.iter_content(chunk_size=1000), q)
-        ).start()
+        # write new file
+        libraryFs.write_file(local_id, file_stream)
 
-        return None
-    except requests.exceptions.HTTPError:
-        raise RecordingNotFound('remote location unreachable')
-    except requests.exceptions.Timeout:
-        raise RecordingNotFound('remote location unreachable')
+        # remove old file if exists
+        if library_model.location_local is not None:
+            libraryFs.remove_file(library_model.location_local)
 
-def read_recording(video_id):
-    library_model = get_library_model(video_id)
+        # update db model with new location
+        library_model.location_local = local_id
+        library_model.save()
+    except Exception as e:
+        # if anything goes wrong, remove_file
+        libraryFs.remove_file(local_id)
+        raise e
+    
+
+def sync_recording(library_model, file_stream):
+    if isinstance(library_model, str):
+        library_model = get_library_model(library_model)
+    
+    threading.Thread(
+        target=do_sync_recording,
+        args=(library_model, file_stream)
+    ).start()
+
+    return None
+
+def read_recording(library_model):
+    if isinstance(library_model, str):
+        library_model = get_library_model(library_model)
+
     try:
         return libraryFs.read_file(library_model.location_local)
     except DoesNotExist:
         raise RecordingNotFound("recording does not exist")
     except FileNotFoundError:
         raise RecordingNotFound("local file not found")
+
+def get_all():
+    return list(Library.select().dicts())
